python/requirements.txt
absl-py==0.7.0
astor==0.7.1
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
imageio==2.4.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.16.1
opencv-python==3.4.2.17
opencv-contrib-python==3.4.2.17
Pillow==5.3.0
plyfile==0.7
protobuf==3.6.1
PyExifTool==0.1.1
pymap3d==1.7.13
pyparsing==2.3.1
pyquaternion==0.9.5
python-dateutil==2.7.5
PyWavelets==1.0.1
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
tensorboard==1.14.0
tensorboardX==1.6
tensorflow-gpu==1.14.0
termcolor==1.1.0
toolz==0.9.0
torch==1.3.1
torchvision==0.4.2
tqdm==4.30.0
Werkzeug==0.14.1
flickrapi==2.4.0
